requirements.txt
Flask==1.1.2
bert-extractive-summarizer==0.7.1
torch
Flask-WTF